refactor(routes): move discovery routes to blueprint
